# docs/swagger.yaml
definitions:
...
  entities.Setting:
    properties:
      setting:
        type: string
      value:
        type: string
      version:
        type: integer
    type: object
...
  managers.SettingSetParam:
    additionalProperties: true
    type: object
  swagger.APIForbiddenError:
    properties:
      message:
        type: string
    type: object
  swagger.APIInternalServerError:
    properties:
      detail:
        type: string
      error:
        type: string
      message:
        type: string
    type: object
  swagger.APIResponse:
    properties:
      data: {}
      status:
        type: string
    type: object
...
  swagger.SettingList:
    properties:
      columns:
        items:
          type: string
        type: array
      data:
        items:
          $ref: '#/definitions/entities.Setting'
        type: array
      total:
        type: integer
    type: object
  swagger.Settings:
    properties:
      settings:
        $ref: '#/definitions/swagger.SettingList'
    type: object
...
info:
  contact: {}
paths:
...
  /setting/general:
    get:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/swagger.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/swagger.Settings'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.APIForbiddenError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.APIInternalServerError'
      tags:
      - Setting
    post:
      consumes:
      - application/json
      parameters:
      - description: ' '
        in: header
        name: Authorization
        required: true
        type: string
      - description: ' '
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/managers.SettingSetParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/swagger.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/swagger.Settings'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.APIForbiddenError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.APIInternalServerError'
      tags:
      - Setting
swagger: "2.0"